ISSUE #4731 - move selectTicketsByModelId to tickets selectors
